Review fixes: merge VirtualKeys tests, fix Logs assertions, tighten AppLogs
- Merge two VirtualKeys test files into one (90 tests, was 60+48 with overlaps)
  - Kept web/src/pages/VirtualKeys/__tests__/VirtualKeys.test.tsx
  - Deleted web/src/pages/__tests__/VirtualKeys.test.tsx
  - Converted old patterns to match new file conventions
- Fix Logs PaginationBar zero-entries: use role-based query instead of regex
- Fix Logs Page Size Change: verify API called with correct per_page param
- Fix Logs Status Badge Muted: use toHaveClass instead of closest+className
- Tighten AppLogs no-source assertion: toBe(1) instead of toBeGreaterThanOrEqual(1)
